guard-platform-CE/core_framework/security_services/algo5/requirements.txt
urllib3==1.13.1
requests==2.9.1
psutil==5.7.3
python-gvm==20.11.0
kafka-python==2.0.2
requests-oauthlib
